requirements.txt
uvicorn~=0.18.3
numpy~=1.22.4
pandas~=1.4.3
fastapi~=0.81.0
python-dateutil~=2.8.2
pydantic~=1.9.2